+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Specify --http-host CLI flag in GetStartCommand to have RPC calls bind to publicIP
 * Migrate AdminAPI endpoints to the new InfoAPI
 * Parse error message on XChain Imports in order to wait for PChain transaction to be accepted
+* Added tests for XChain endpoints
 
 # 0.3.0
 * Removed startPortRange and endPortRange CLI args